test(places): fix + consolidate PropertyInfoForm tests under a stub router (#200)
The #200 useFetcher addition made PropertyInfoForm require a data-router context,
breaking the pre-existing property-info-form.test.tsx (which rendered it bare).
Wrap all renders in a createRoutesStub router, fold the autofill-button tests
into the same file, and drop the temporary duplicate PropertyInfoForm.test.tsx.
