+++ app/components/editor/property-info-form.test.tsx
@@ -1,12 +1,20 @@
 import { useState } from "react";
-import { render, fireEvent } from "@testing-library/react";
+import { render, fireEvent, screen, waitFor } from "@testing-library/react";
+import { createRoutesStub } from "react-router";
 import { PropertyInfoForm } from "~/components/editor/PropertyInfoForm";
 
-// PropertyInfoForm takes plain callback props (onSave/onCommit) and no
-// fetcher, so it renders directly without a router (unlike CompliancePanel).
+// PropertyInfoForm now uses useFetcher (the #200 "Fetch property details"
+// button), so it must render inside a data-router context. `renderPF` wraps any
+// element in a minimal stub router; an optional `action` powers the autofill
+// fetcher for the tests that click the button. The field/commit tests below
+// never click it, so their fetcher stays idle and no action is needed.
 // `onSave` fires optimistically on every change with the RAW value typed;
 // `onCommit` fires the durable save with a FULL coerced snapshot of every field
 // — on blur for text/number/date, on change for select/boolean.
+function renderPF(ui: React.ReactElement, action?: () => unknown) {
+  const Stub = createRoutesStub([{ path: "/", Component: () => ui, action }]);
+  return render(<Stub />);
+}
 
 function labelInput(container: HTMLElement, labelText: string): HTMLInputElement | null {
   const spans = Array.from(container.querySelectorAll("label span span")) as HTMLElement[];
@@ -33,7 +41,7 @@ function Harness({ initial, onCommit }: { initial: Record<string, unknown>; onCo
 
 describe("PropertyInfoForm field sets", () => {
   it("renders bedrooms/bathrooms plus unit + county for a residential inspection", () => {
-    const { getByText } = render(
+    const { getByText } = renderPF(
       <PropertyInfoForm inspection={{ propertyType: "residential" }} />,
     );
     expect(getByText("Bedrooms")).toBeTruthy();
@@ -45,7 +53,7 @@ describe("PropertyInfoForm field sets", () => {
   });
 
   it("omits bedrooms/bathrooms + unit but shows Building Area and County for a commercial inspection", () => {
-    const { queryByText, getByText } = render(
+    const { queryByText, getByText } = renderPF(
       <PropertyInfoForm inspection={{ propertyType: "commercial" }} />,
     );
     expect(queryByText("Bedrooms")).toBeNull();
@@ -59,7 +67,7 @@ describe("PropertyInfoForm field sets", () => {
   });
 
   it("renders lotSize (Lot Size)", () => {
-    const { getByText } = render(
+    const { getByText } = renderPF(
       <PropertyInfoForm inspection={{ propertyType: "residential" }} />,
     );
     expect(getByText("Lot Size")).toBeTruthy();
@@ -69,7 +77,7 @@ describe("PropertyInfoForm field sets", () => {
 describe("PropertyInfoForm raw-string typing (Finding 4)", () => {
   it("keeps a fractional Bathrooms value ('2.5') verbatim — onChange is the raw string, not coerced to 2", () => {
     const onSave = vi.fn();
-    const { container } = render(
+    const { container } = renderPF(
       <PropertyInfoForm inspection={{ propertyType: "residential" }} onSave={onSave} />,
     );
     const input = labelInput(container, "Bathrooms")!;
@@ -81,7 +89,7 @@ describe("PropertyInfoForm raw-string typing (Finding 4)", () => {
 
   it("does not clear the input on the transient mid-decimal '2.' keystroke", () => {
     const onCommit = vi.fn();
-    const { container } = render(<Harness initial={{ propertyType: "residential" }} onCommit={onCommit} />);
+    const { container } = renderPF(<Harness initial={{ propertyType: "residential" }} onCommit={onCommit} />);
     const input = labelInput(container, "Bathrooms")!;
     fireEvent.change(input, { target: { value: "2." } });
     // Controlled value retains "2." — a coerce here would snap it to "2" or "".
@@ -94,7 +102,7 @@ describe("PropertyInfoForm raw-string typing (Finding 4)", () => {
 describe("PropertyInfoForm full-snapshot commit (Finding 1)", () => {
   it("commits the WHOLE facts object on blur with the changed field coerced and empty fields → null", () => {
     const onCommit = vi.fn();
-    const { container } = render(<Harness initial={{ propertyType: "residential" }} onCommit={onCommit} />);
+    const { container } = renderPF(<Harness initial={{ propertyType: "residential" }} onCommit={onCommit} />);
     const input = labelInput(container, "Year Built")!;
     fireEvent.change(input, { target: { value: "1990" } });
     // onChange must NOT commit (no keystroke persistence).
@@ -115,7 +123,7 @@ describe("PropertyInfoForm full-snapshot commit (Finding 1)", () => {
 
   it("commits a fractional bathrooms value coerced to a Number within the full object", () => {
     const onCommit = vi.fn();
-    const { container } = render(<Harness initial={{ propertyType: "residential" }} onCommit={onCommit} />);
+    const { container } = renderPF(<Harness initial={{ propertyType: "residential" }} onCommit={onCommit} />);
     const input = labelInput(container, "Bathrooms")!;
     fireEvent.change(input, { target: { value: "2.5" } });
     fireEvent.blur(input);
@@ -125,7 +133,7 @@ describe("PropertyInfoForm full-snapshot commit (Finding 1)", () => {
 
   it("preserves already-entered sibling fields when a second field commits (superset PATCH)", () => {
     const onCommit = vi.fn();
-    const { container } = render(<Harness initial={{ propertyType: "residential" }} onCommit={onCommit} />);
+    const { container } = renderPF(<Harness initial={{ propertyType: "residential" }} onCommit={onCommit} />);
     const year = labelInput(container, "Year Built")!;
     fireEvent.change(year, { target: { value: "1990" } });
     fireEvent.blur(year);
@@ -140,7 +148,7 @@ describe("PropertyInfoForm full-snapshot commit (Finding 1)", () => {
 
   it("commits null when a number field is cleared to empty on blur", () => {
     const onCommit = vi.fn();
-    const { container } = render(<Harness initial={{ propertyType: "residential", yearBuilt: 1990 }} onCommit={onCommit} />);
+    const { container } = renderPF(<Harness initial={{ propertyType: "residential", yearBuilt: 1990 }} onCommit={onCommit} />);
     const input = labelInput(container, "Year Built")!;
     fireEvent.change(input, { target: { value: "" } });
     fireEvent.blur(input);
@@ -150,7 +158,7 @@ describe("PropertyInfoForm full-snapshot commit (Finding 1)", () => {
 
   it("commits null when a text field (lotSize) is blurred empty", () => {
     const onCommit = vi.fn();
-    const { container } = render(<Harness initial={{ propertyType: "residential", lotSize: "old" }} onCommit={onCommit} />);
+    const { container } = renderPF(<Harness initial={{ propertyType: "residential", lotSize: "old" }} onCommit={onCommit} />);
     const input = labelInput(container, "Lot Size")!;
     expect(input).toBeTruthy();
     fireEvent.change(input, { target: { value: "" } });
@@ -161,7 +169,7 @@ describe("PropertyInfoForm full-snapshot commit (Finding 1)", () => {
 
   it("commits a text field (lotSize) value on blur", () => {
     const onCommit = vi.fn();
-    const { container } = render(<Harness initial={{ propertyType: "residential" }} onCommit={onCommit} />);
+    const { container } = renderPF(<Harness initial={{ propertyType: "residential" }} onCommit={onCommit} />);
     const input = labelInput(container, "Lot Size")!;
     fireEvent.change(input, { target: { value: "0.25 acres" } });
     fireEvent.blur(input);
@@ -171,7 +179,7 @@ describe("PropertyInfoForm full-snapshot commit (Finding 1)", () => {
 
   it("still fires onSave optimistically on change with the raw string", () => {
     const onSave = vi.fn();
-    const { container } = render(
+    const { container } = renderPF(
       <PropertyInfoForm inspection={{ propertyType: "residential" }} onSave={onSave} />,
     );
     const input = labelInput(container, "Year Built")!;
@@ -193,7 +201,7 @@ describe("PropertyInfoForm commercial preset metadata split", () => {
   ];
 
   it("renders subtype-preset fields passed via templateFields", () => {
-    const { getByText } = render(
+    const { getByText } = renderPF(
       <PropertyInfoForm inspection={{ propertyType: "commercial" }} templateFields={officePreset} />,
     );
     expect(getByText("Net rentable area")).toBeTruthy();
@@ -202,7 +210,7 @@ describe("PropertyInfoForm commercial preset metadata split", () => {
 
   it("splits non-dedicated preset fields into a metadata envelope on commit", () => {
     const onCommit = vi.fn();
-    const { container } = render(
+    const { container } = renderPF(
       <PropertyInfoForm
         inspection={{ propertyType: "commercial", yearBuilt: 1998, nra: 42000 }}
         templateFields={officePreset}
@@ -221,7 +229,7 @@ describe("PropertyInfoForm commercial preset metadata split", () => {
       { id: "yearBuilt", label: "Year built", type: "number" as const, group: "Identity" },
       { id: "sqft", label: "Building area", type: "number" as const, group: "Physical" },
     ];
-    const { container } = render(
+    const { container } = renderPF(
       <PropertyInfoForm
         inspection={{ propertyType: "commercial", yearBuilt: 2001 }}
         templateFields={dedicatedOnly}
@@ -238,7 +246,7 @@ describe("PropertyInfoForm commercial preset metadata split", () => {
 describe("PropertyInfoForm select commit", () => {
   it("commits foundationType on change (discrete select) with the full facts object", () => {
     const onCommit = vi.fn();
-    const { container } = render(
+    const { container } = renderPF(
       <PropertyInfoForm inspection={{ propertyType: "residential" }} onCommit={onCommit} />,
     );
     const select = container.querySelector("select") as HTMLSelectElement;
@@ -254,5 +262,53 @@ describe("PropertyInfoForm select commit", () => {
       unit: null,
       county: null,
     });
+  });
+});
+
+// #200 — "Fetch property details" autofill button. Relays the address to the
+// editor BFF action; fills ONLY empty fields (never clobbers manual entry) and
+// commits the merged snapshot; degrades with a clear message when unconfigured.
+describe("PropertyInfoForm — Fetch property details (#200)", () => {
+  function mountAutofill(
+    actionResult: unknown,
+    handlers: { onSave?: (id: string, v: unknown) => void; onCommit?: (f: Record<string, unknown>) => void },
+  ) {
+    const inspection: Record<string, unknown> = {
+      propertyType: "single_family",
+      propertyAddress: "123 Main St, Austin, TX 78701",
+      bedrooms: "4", // already set — must NOT be overwritten
+      // yearBuilt / sqft / etc. absent = empty = fillable
+    };
+    return renderPF(
+      <PropertyInfoForm inspection={inspection} onSave={handlers.onSave} onCommit={handlers.onCommit} />,
+      () => actionResult,
+    );
+  }
+
+  it("fills only empty fields and never clobbers a value the inspector already set", async () => {
+    const onSave = vi.fn();
+    const onCommit = vi.fn();
+    mountAutofill(
+      { intent: "autofill-property-facts", facts: { yearBuilt: 1990, bedrooms: 3 }, reason: null },
+      { onSave, onCommit },
+    );
+
+    fireEvent.click(screen.getByRole("button", { name: /fetch property details/i }));
+
+    await waitFor(() => expect(onSave).toHaveBeenCalledWith("yearBuilt", 1990));
+    expect(onSave).not.toHaveBeenCalledWith("bedrooms", 3);
+    expect(onCommit).toHaveBeenCalledWith(expect.objectContaining({ yearBuilt: 1990, bedrooms: 4 }));
+    await waitFor(() => expect(screen.getByText(/filled 1 field/i)).toBeTruthy());
+  });
+
+  it("shows an 'unconfigured' message and commits nothing when the provider key is unset", async () => {
+    const onSave = vi.fn();
+    const onCommit = vi.fn();
+    mountAutofill({ intent: "autofill-property-facts", facts: null, reason: "NO_API_KEY" }, { onSave, onCommit });
+
+    fireEvent.click(screen.getByRole("button", { name: /fetch property details/i }));
+
+    await waitFor(() => expect(screen.getByText(/isn.t configured/i)).toBeTruthy());
+    expect(onCommit).not.toHaveBeenCalled();
   });
 });
